requirements.txt
albumentations~=0.5.1
einops~=0.3.0
future==0.18.2
fvcore==0.1.5.post20221221
h5py==3.1.0
joblib~=1.2.0
kornia~=0.7.1
loguru~=0.5.3
matplotlib~=3.7.1
opencv-python~=4.5.5.62
Pillow~=9.4.0
pytorch-lightning==1.9.0
scikit-image==0.19.1
scikit-learn~=1.2.2
tqdm~=4.65.0
yacs==0.1.8
torchmetrics==0.11.0
gdown
tensorboardX==2.6
timm~=0.6.13
thop==0.1.1.post2310181704
numpy~=1.24.2
torch~=1.10.0+cu102
pyyaml~=5.4.1
lightning~=1.9.0